feat: introduces zod to api
