+++ packages/swapkit/api/src/thorswapApi/types.ts
@@ -1,4 +1,5 @@
 import type { Chain, LedgerErrorCode, QuoteMode } from "@swapkit/helpers";
+import { z } from "zod";
 
 export type FeeType = "AFFILIATE" | "LIQUIDITY" | "INBOUND" | "OUTBOUND" | "NETWORK";
 
@@ -48,8 +49,6 @@ export type QuoteRouteV2 = {
   targetAddress?: string;
   providers: string[];
 };
-
-export type QuoteResponse = { quoteId: string; routes: QuoteRoute[] };
 
 export type LoansParams = { address: string; asset: string };
 
@@ -224,27 +223,29 @@ export type RepayStreamingSwap = {
   estimatedTime?: number;
 };
 
-export type RepayResponse = {
-  inboundAddress: string;
-  inboundConfirmationBlocks: number;
-  inboundConfirmationSeconds: number;
-  outboundDelayBlocks: number;
-  outboundDelaySeconds: number;
-  fees: { asset: string; liquidity: string; totalBps: number };
-  expiry: number;
-  warning?: string;
-  notes?: string;
-  dustThreshold: string;
-  memo: string;
-  expectedAmountOut: string;
-  expectedCollateralWithdrawn: string;
-  expectedDebtRepaid: string;
-  collateralCurrent: string;
-  repayAssetAmount: string;
-  repayAssetAmountUSD: string;
-  streamingSwap?: RepayStreamingSwap;
-  estimatedTime?: number;
-};
+export type RepayResponse =
+  | ApiV1Error
+  | {
+      inboundAddress: string;
+      inboundConfirmationBlocks: number;
+      inboundConfirmationSeconds: number;
+      outboundDelayBlocks: number;
+      outboundDelaySeconds: number;
+      fees: { asset: string; liquidity: string; totalBps: number };
+      expiry: number;
+      warning?: string;
+      notes?: string;
+      dustThreshold: string;
+      memo: string;
+      expectedAmountOut: string;
+      expectedCollateralWithdrawn: string;
+      expectedDebtRepaid: string;
+      collateralCurrent: string;
+      repayAssetAmount: string;
+      repayAssetAmountUSD: string;
+      streamingSwap?: RepayStreamingSwap;
+      estimatedTime?: number;
+    };
 
 export type BorrowParams = {
   assetIn: string;
@@ -278,29 +279,33 @@ export type BorrowStreamingSwap = {
   memo: string;
 };
 
-export type BorrowResponse = {
-  amountIn: string;
-  amountOut: string;
-  amountOutMin: string;
-  calldata: BorrowCalldata;
-  complete: boolean;
-  estimatedTime: number;
-  expectedCollateralDeposited: string;
-  expectedDebtIssued: string;
-  expectedOutput: string;
-  expectedOutputMaxSlippage: string;
-  expectedOutputMaxSlippageUSD: string;
-  expectedOutputUSD: string;
-  fees: QuoteRoute["fees"];
-  fromAsset: string;
-  memo: string;
-  recipientAddress: string;
-  route: { meta: { thornodeMeta: { inboundConfirmationSeconds: number; outboundDelay: number } } };
-  streamingSwap?: BorrowStreamingSwap;
-  swaps: QuoteRoute["swaps"];
-  targetAddress: string;
-  toAsset: string;
-};
+export type BorrowResponse =
+  | ApiV1Error
+  | {
+      amountIn: string;
+      amountOut: string;
+      amountOutMin: string;
+      calldata: BorrowCalldata;
+      complete: boolean;
+      estimatedTime: number;
+      expectedCollateralDeposited: string;
+      expectedDebtIssued: string;
+      expectedOutput: string;
+      expectedOutputMaxSlippage: string;
+      expectedOutputMaxSlippageUSD: string;
+      expectedOutputUSD: string;
+      fees: QuoteRoute["fees"];
+      fromAsset: string;
+      memo: string;
+      recipientAddress: string;
+      route: {
+        meta: { thornodeMeta: { inboundConfirmationSeconds: number; outboundDelay: number } };
+      };
+      streamingSwap?: BorrowStreamingSwap;
+      swaps: QuoteRoute["swaps"];
+      targetAddress: string;
+      toAsset: string;
+    };
 
 export type EVMTransactionDetails = {
   approvalSpender?: string;
@@ -503,3 +508,124 @@ type Meta = {
     warning: string;
   };
 };
+
+export enum ERROR_CODE {
+  INVALID_INPUT_PARAMETERS = "1000",
+  UNKNOWN_PROVIDERS = "1001",
+  CANNOT_FIND_INBOUND_ADDRESS = "1002",
+  NO_INBOUND_ADDRESSES = "1003",
+  CHAIN_HALTED_OR_UNSUPPORTED = "1004",
+  MISSING_INPUT_PARAMETER = "1005",
+  INVALID_TYPE_GENERIC = "1100",
+  INVALID_NUMBER_STRING = "1101",
+  INVALID_NUMBER = "1102",
+  INVALID_BOOLEAN = "1103",
+  INVALID_OBJECT = "1104",
+  INVALID_ARRAY = "1105",
+  SELL_AMOUNT_MUST_BE_POSITIVE_INTEGER = "2000",
+  SELL_BUY_ASSETS_ARE_THE_SAME = "2001",
+  MISSING_SOURCE_ADDRESS_FOR_SYNTH = "2002",
+  AFF_ADDRESS_AND_BPS_OR_NEITHER = "2003",
+  AFF_ADDRESS_TOO_LONG = "2004",
+  AFF_BPS_INTEGER_MAX_500 = "2005",
+  SOURCE_ADDRESS_INVALID_FOR_SELL_CHAIN = "2006",
+  DESTINATION_ADDRESS_INVALID_FOR_BUY_CHAIN = "2007",
+  PREFERRED_PROVIDER_NOT_SUPPORTED = "2008",
+  DESTINATION_ADDRESS_SMART_CONTRACT = "2009",
+  BUY_AMOUNT_MUST_BE_POSITIVE_INTEGER = "2010",
+  SOURCE_ADDRESS_SMART_CONTRACT = "2011",
+  SWAP_AMOUNT_TOO_LOW = "2012",
+  INVALID_PROVIDER = "2100",
+  MISSING_CROSS_CHAIN_PROVIDER = "2101",
+  MISSING_AVAX_PROVIDER = "2102",
+  MISSING_BSC_PROVIDER = "2103",
+  MISSING_ETH_PROVIDER = "2104",
+  INVALID_PROVIDER_FOR_SWAP_OUT = "2105",
+  MISSING_ARB_PROVIDER = "2106",
+  INVALID_CHAIN = "2200",
+  INVALID_ASSET = "2201",
+  INVALID_ASSET_IDENTIFIER = "2202",
+  UNSUPPORTED_CHAIN = "2204",
+  UNSUPPORTED_ASSET = "2203",
+  UNSUPPORTED_ASSET_FOR_SWAPOUT = "2205",
+  INVALID_SOURCE_ADDRESS = "2300",
+  INVALID_DESTINATION_ADDRESS = "2301",
+  THORNODE_QUOTE_GENERIC_ERROR = "3000",
+  NOT_ENOUGH_SYNTH_BALANCE = "3001",
+  SYNTH_MINTING_CAP_REACHED = "3002",
+  INVALID_QUOTE_MODE = "4000",
+  NO_QUOTES = "4001",
+  SERVICE_UNAVAILABLE_GENERIC = "5000",
+  MISSING_GAS_DATA_GENERIC = "5100",
+  MISSING_TOKEN_INFO_GENERIC = "5200",
+  CANT_FIND_TOKEN_LIST = "5201",
+  NO_PRICE = "5202",
+  PRICE_IS_STALE = "5203",
+  ADDRESS_NOT_WHITELISTED = "6000",
+  ADDRESS_ALREADY_CLAIMED = "6001",
+  TEMPORARY_ERROR = "9999",
+}
+
+export enum ERROR_MODULE {
+  HEALTH_CONTROLLER = "1000",
+  LIQUIDITY_CONTROLLER = "1001",
+  PROVIDER_CONTROLLER = "1002",
+  QUOTE_CONTROLLER = "1003",
+  SWAP_CONTROLLER = "1004",
+  UTIL_CONTROLLER = "1005",
+  AIRDROP_CONTROLLER = "1006",
+  PROVIDER = "2000",
+  ASSET = "2001",
+  TOKEN_LIST = "2002",
+  QUOTE = "2100",
+  QUOTE_TXN_DETAILS = "2101",
+  THORCHAIN_PROVIDER = "3000",
+  UNISWAPV2_ETH_PROVIDER = "3001",
+  UNISWAPV3_ETH_PROVIDER = "3002",
+  SUSHISWAP_ETH_PROVIDER = "3003",
+  PANCAKESWAP_BSC_PROVIDER = "3004",
+  PANCAKESWAP_ETH_PROVIDER = "3005",
+  ONEINCH_ETH_PROVIDER = "3006",
+  ONEINCH_BSC_PROVIDER = "3007",
+  ONEINCH_AVAX_PROVIDER = "3008",
+  ZEROX_ETH_PROVIDER = "3009",
+  WOOFI_AVAX_PROVIDER = "3010",
+  PANGOLIN_AVAX_PROVIDER = "3011",
+  TRADERJOE_AVAX_PROVIDER = "3012",
+  KYBER_ETH_PROVIDER = "3013",
+  KYBER_AVAX_PROVIDER = "3014",
+  WOOFI_BSC_PROVIDER = "3015",
+  STARGATE_PROVIDER = "3016",
+  PROVIDER_UTIL = "4000",
+  TXN_DETAILS = "5000",
+  AIRDROP_UTIL = "6000",
+}
+
+export declare enum ERROR_TYPE {
+  VALIDATION_ERROR = "VALIDATION_ERROR",
+  REQUEST_PARAMETER_ERROR = "REQUEST_PARAMETER_ERROR",
+  RESPONSE_PARSING_ERROR = "RESPONSE_PARSING_ERROR",
+  UNSUPPORTED = "UNSUPPORTED",
+  NOT_IMPLEMENTED = "NOT_IMPLEMENTED",
+  INCOMPATIBLE_ASSETS_OPERATIONS = "INCOMPATIBLE_ASSETS_OPERATIONS",
+  SERVICE_UNAVAILABLE = "SERVICE_UNAVAILABLE",
+  DOWN_FOR_MAINTENANCE = "DOWN_FOR_MAINTENANCE",
+  MISSING_INBOUND_INFO = "MISSING_INBOUND_INFO",
+  QUOTE_FETCHING_ERROR = "QUOTE_FETCHING_ERROR",
+  AIRDROP_ERROR = "AIRDROP_ERROR",
+  UNHANDLED_ERROR = "UNHANDLED_ERROR",
+}
+
+export const ApiV1ErrorSchema = z.object({
+  status: z.number(),
+  type: z.nativeEnum(ERROR_TYPE),
+  code: z.nativeEnum(ERROR_CODE),
+  module: z.nativeEnum(ERROR_MODULE),
+  complete: z.string(),
+  identifier: z.string(),
+  message: z.string(),
+});
+
+export type ApiV1Error = z.infer<typeof ApiV1ErrorSchema>;
+
+export type QuoteResponse = { quoteId: string; routes: QuoteRoute[] } | ApiV1Error;
