+++ packages/swapkit/helpers/src/modules/__tests__/requestClient.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, test } from "bun:test";
 
-import { RequestClientWithErrors } from "../requestClient";
+import { RequestClient } from "../requestClient";
 
 describe("ApiV1 error handling", () => {
   test("Expected error", async () => {
-    const expectedError = await RequestClientWithErrors.get(
+    const expectedError = await RequestClient.get(
       "https://api.thorswap.finance/aggregator/tokens/quote?sellAsset=ETH.00-0X881BA05DE1E78F549CC63A8F6CABB1D4AD32250D&sellAmount=1.12&buyAsset=ETH.ETH&senderAddress=0x494447b317d2ee41d4c02600edb7c2193b2c9085&recipientAddress=0x494447b317d2ee41d4c02600edb7c2193b2c9085&slippage=3",
     );
