+++ playgrounds/nextjs/package.json
@@ -59,7 +59,7 @@
     "sonner": "1.4.41",
     "tailwind-merge": "2.2.2",
     "tailwindcss-animate": "1.0.7",
-    "zod": "3.22.4"
+    "zod": "3.23.8"
   },
   "devDependencies": {
     "@swc-jotai/debug-label": "0.1.0",
